Rename src/macos.rs -> src/bsd.rs
